refactor(api-service): split ManagedAgentService into focused services fixes NV-7792
Extract ManagedAgentProviderFactory (LRU-cached provider resolution) and
ManagedAgentEventHandler (webhook routing, error recovery, card building)
from the monolithic ManagedAgentService, leaving it as a slim facade for
dispatch, confirmToolApproval, and handleWebhook.

Unify plan delivery through HandleAgentReply via a new `plan` command
field so HandlePlanProgress (renamed from HandleToolProgress) delegates
to the single outbound delivery path instead of calling ChatSdkService
directly. Add `awaiting-approval` action for plan card state during
tool confirmation flows.

Move parseToolApprovalActionId and isLinkButtonActionId to the event
handler as their canonical home.
