+++ apps/api/src/app/agents/usecases/handle-agent-reply/handle-agent-reply.command.ts
@@ -1,4 +1,5 @@
 import type { Signal } from '@novu/framework';
+import type { PlanModel } from 'chat';
 import { Type } from 'class-transformer';
 import { IsArray, IsNotEmpty, IsObject, IsOptional, IsString, ValidateNested } from 'class-validator';
 import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
@@ -40,4 +41,8 @@ export class HandleAgentReplyCommand extends EnvironmentWithUserCommand {
   @ValidateNested({ each: true })
   @Type(() => AddReactionPayloadDto)
   addReactions?: AddReactionPayloadDto[];
+
+  @IsOptional()
+  @IsObject()
+  plan?: { model: PlanModel; messageId?: string };
 }
